add the order, review... stuff

+++ src/main/java/com/ouss/ecom/dao/UserRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 
-public interface UserRepo extends JpaRepository<AppUser, Integer> {
+public interface UserRepo extends JpaRepository<AppUser, Long> {
     Optional<AppUser> findByEmail(String email);
 
 }
